config: move temp parsing to temps
